Refactor project into modular task packages
