+++ src/ablations/__init__.py
@@ -1,0 +1,6 @@
+"""Tools for applying structural ablations."""
+
+from .apply import apply_ablation
+from .config import AblationConfig
+
+__all__ = ["AblationConfig", "apply_ablation"]
