+++ src/inference/__init__.py
@@ -1,0 +1,6 @@
+"""Inference helpers."""
+
+from .cli import main
+from .greedy_decoder import GreedyDecoder
+
+__all__ = ["GreedyDecoder", "main"]
