+++ src/models/__init__.py
@@ -1,0 +1,12 @@
+"""Model components for toy-language transformer experiments."""
+
+from .config import TransformerConfig
+from .toy_transformer import ToyTransformer
+from .utils import build_padding_mask, generate_causal_mask
+
+__all__ = [
+    "TransformerConfig",
+    "ToyTransformer",
+    "build_padding_mask",
+    "generate_causal_mask",
+]
