+++ src/models/positional_encoding.py
@@ -1,0 +1,25 @@
+from __future__ import annotations
+
+import math
+
+import torch
+from torch import Tensor, nn
+
+
+class PositionalEncoding(nn.Module):
+    """Classic sinusoidal positional encoding with dropout."""
+
+    def __init__(self, d_model: int, dropout: float = 0.1, max_len: int = 5000) -> None:
+        super().__init__()
+        self.dropout = nn.Dropout(dropout)
+        pe = torch.zeros(max_len, d_model)
+        position = torch.arange(0, max_len, dtype=torch.float).unsqueeze(1)
+        div_term = torch.exp(torch.arange(0, d_model, 2).float() * (-math.log(10000.0) / d_model))
+        pe[:, 0::2] = torch.sin(position * div_term)
+        pe[:, 1::2] = torch.cos(position * div_term)
+        pe = pe.unsqueeze(0)
+        self.register_buffer("pe", pe)
+
+    def forward(self, x: Tensor) -> Tensor:
+        x = x + self.pe[:, : x.size(1)]
+        return self.dropout(x)
