+++ src/training/__init__.py
@@ -1,0 +1,6 @@
+"""Training utilities."""
+
+from .cli import main
+from .loop import run_epoch, train_loop
+
+__all__ = ["main", "run_epoch", "train_loop"]
