+++ src/training/config.py
@@ -1,0 +1,11 @@
+from __future__ import annotations
+
+from pathlib import Path
+from typing import Dict
+
+import yaml
+
+
+def load_config(path: str | Path) -> Dict[str, object]:
+    with Path(path).open("r", encoding="utf-8") as f:
+        return yaml.safe_load(f)
