+++ src/vocabulary/__init__.py
@@ -1,0 +1,7 @@
+"""Vocabulary utilities."""
+
+from .builder import build_vocabulary_from_files
+from .constants import SPECIAL_TOKENS
+from .vocabulary import Vocabulary
+
+__all__ = ["Vocabulary", "SPECIAL_TOKENS", "build_vocabulary_from_files"]
