+++ src/vocabulary/constants.py
@@ -1,0 +1,7 @@
+SPECIAL_TOKENS = {
+    "pad": "<PAD>",
+    "bos": "<BOS>",
+    "eos": "<EOS>",
+    "unk": "<UNK>",
+    "span": "<SPAN>",
+}
